OSIS-5832 moved training builder
