+++ ddd/logic/formation_catalogue/builder/training_builder.py
@@ -23,6 +23,8 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
+import attr
+
 from base.models.enums.academic_type import AcademicTypes
 from base.models.enums.active_status import ActiveStatusEnum
 from base.models.enums.activity_presence import ActivityPresence
@@ -34,6 +36,7 @@ from base.models.enums.funding_codes import FundingCodes
 from base.models.enums.internship_presence import InternshipPresence
 from base.models.enums.rate_code import RateCode
 from base.models.enums.schedule_type import ScheduleTypeEnum
+from base.models.utils import utils
 from ddd.logic.formation_catalogue.domain.model._first_year_bachelor import FirstYearBachelor, FirstYearBachelorIdentity
 from ddd.logic.formation_catalogue.domain.model.bachelor import Bachelor
 from ddd.logic.formation_catalogue.dtos import TrainingDto, BachelorDto
@@ -49,8 +52,12 @@ from education_group.ddd.domain._isced_domain import IscedDomain, IscedDomainIde
 from education_group.ddd.domain._language import Language
 from education_group.ddd.domain._study_domain import StudyDomain, StudyDomainIdentity
 from education_group.ddd.domain._titles import Titles
+from education_group.ddd.domain.exception import TrainingNotFoundException
 from education_group.ddd.domain.training import Training, TrainingIdentity, TrainingIdentityThroughYears
-from education_group.ddd.domain._entity import Entity as EntityValueObject
+from education_group.ddd.domain._entity import Entity as EntityValueObject, Entity
+from education_group.ddd.validators.validators_by_business_action import CopyTrainingValidatorList, \
+    CreateTrainingValidatorList
+from osis_common.ddd import interface
 from osis_common.ddd.interface import RootEntityBuilder
 
 
@@ -197,3 +204,139 @@ class TrainingBuilder(RootEntityBuilder):
                 **datas
             )
         return Training(**datas)
+
+    def copy_to_next_year(self, training_from: 'Training', training_repository: 'TrainingRepository') -> 'Training':
+        identity_next_year = TrainingIdentity(acronym=training_from.acronym, year=training_from.year + 1)
+        CopyTrainingValidatorList(training_from).validate()
+        try:
+            training_next_year = training_repository.get(identity_next_year)
+            training_next_year.update_from_other_training(training_from)
+        except TrainingNotFoundException:
+            training_next_year = attr.evolve(
+                training_from,
+                entity_identity=identity_next_year,
+                entity_id=identity_next_year,
+            )
+        return training_next_year
+
+    def copy_aims_to_next_year(
+            self,
+            training_from: 'Training',
+            training_repository: 'TrainingRepository'
+    ) -> 'Training':
+        identity_next_year = TrainingIdentity(acronym=training_from.acronym, year=training_from.year + 1)
+        training_next_year = training_repository.get(identity_next_year)
+        training_next_year.update_aims_from_other_training(training_from)
+        return training_next_year
+
+    def create_training(self, command: 'CreateTrainingCommand') -> 'Training':
+        training_identity = TrainingIdentity(command.abbreviated_title, command.year)
+
+        secondary_domains = []
+        for dom in command.secondary_domains:
+            secondary_domains.append(
+                StudyDomain(
+                    entity_id=StudyDomainIdentity(dom[0], dom[1]),
+                    domain_name=None,
+                )
+            )
+        datas = {
+            'entity_identity': training_identity,
+            'entity_id': training_identity,
+            'code': command.code,
+            'type': utils.get_enum_from_str(command.type, TrainingType),
+            'credits': command.credits,
+            'schedule_type': utils.get_enum_from_str(command.schedule_type, ScheduleTypeEnum),
+            'duration': command.duration,
+            'duration_unit': utils.get_enum_from_str(command.duration_unit, DurationUnitsEnum),
+            'start_year': command.start_year,
+            'titles': Titles(
+                title_fr=command.title_fr,
+                partial_title_fr=command.partial_title_fr,
+                title_en=command.title_en,
+                partial_title_en=command.partial_title_en,
+            ),
+            'status': utils.get_enum_from_str(command.status, ActiveStatusEnum),
+            'keywords': command.keywords,
+            'internship_presence': utils.get_enum_from_str(command.internship_presence, InternshipPresence),
+            'is_enrollment_enabled': command.is_enrollment_enabled,
+            'has_online_re_registration': command.has_online_re_registration,
+            'has_partial_deliberation': command.has_partial_deliberation,
+            'has_admission_exam': command.has_admission_exam,
+            'has_dissertation': command.has_dissertation,
+            'produce_university_certificate': command.produce_university_certificate,
+            'decree_category': utils.get_enum_from_str(command.decree_category, DecreeCategories),
+            'rate_code': utils.get_enum_from_str(command.rate_code, RateCode),
+            'main_language': Language(command.main_language) if command.main_language else None,
+            'english_activities': utils.get_enum_from_str(command.english_activities, ActivityPresence),
+            'other_language_activities': utils.get_enum_from_str(
+                command.other_language_activities, ActivityPresence
+            ),
+            'internal_comment': command.internal_comment,
+            'main_domain': StudyDomain(
+                entity_id=StudyDomainIdentity(decree_name=command.main_domain_decree, code=command.main_domain_code),
+                domain_name=None,
+            ) if command.main_domain_code else None,
+            'secondary_domains': secondary_domains,
+            'isced_domain': IscedDomain(
+                entity_id=IscedDomainIdentity(command.isced_domain_code),
+                title_fr=None,
+                title_en=None,
+            ) if command.isced_domain_code else None,
+            'management_entity': Entity(acronym=command.management_entity_acronym),
+            'administration_entity': Entity(acronym=command.administration_entity_acronym),
+            'end_year': command.end_year,
+            'enrollment_campus': Campus(
+                name=command.enrollment_campus_name,
+                university_name=command.enrollment_campus_organization_name,
+            ),
+            'other_campus_activities': utils.get_enum_from_str(command.other_campus_activities, ActivityPresence),
+            'funding': Funding(
+                can_be_funded=command.can_be_funded,
+                funding_orientation=FundingCodes[
+                    command.funding_orientation
+                ] if command.funding_orientation else None,
+                can_be_international_funded=command.can_be_international_funded,
+                international_funding_orientation=FundingCodes[
+                    command.international_funding_orientation
+                ] if command.international_funding_orientation else None,
+            ),
+            'hops': HOPS(
+                ares_code=command.ares_code,
+                ares_graca=command.ares_graca,
+                ares_authorization=command.ares_authorization,
+            ),
+            'co_graduation': CoGraduation(
+                code_inter_cfb=command.code_inter_cfb,
+                coefficient=command.coefficient,
+            ),
+            'academic_type': utils.get_enum_from_str(command.academic_type, AcademicTypes),
+            'diploma': Diploma(
+                aims=None,
+                leads_to_diploma=command.leads_to_diploma,
+                printing_title=command.printing_title,
+                professional_title=command.professional_title,
+            ),
+        }
+
+        # todo REMOVE cyclic import
+        if datas['type'] == TrainingType.BACHELOR:
+            from ddd.logic.formation_catalogue.domain.model.bachelor import Bachelor
+            from ddd.logic.formation_catalogue.domain.model._first_year_bachelor import FirstYearBachelor, \
+                FirstYearBachelorIdentity
+
+            training = Bachelor(first_year_bachelor=FirstYearBachelor(entity_id=FirstYearBachelorIdentity(), administration_entity=None), **datas)
+        else:
+            training = Training(**datas)
+        CreateTrainingValidatorList(training).validate()
+        return training
+
+    def _get_enum_from_str(self, value: str, enum_class):
+        if not value:
+            return None
+        try:
+            return enum_class[value]
+        except ValueError:
+            raise interface.BusinessException(
+                "Invalid enum choice (value={}, enumeration_class={})".format(value, enum_class)
+            )
